+++ R/zzz.R
@@ -1,0 +1,5 @@
+.onLoad <- function(libname, pkgname) {
+  theme_set(theme_ptt())
+  packageStartupMessage("theme_ptt set for ggplot2")
+  invisible()
+}