+++ NEWS.md
@@ -4,6 +4,8 @@ ggptt 0.2.0
 * `the_legend_bot for legend.position = "bottom"`
 * `ptt_pal()` color PTT palette. 6 colours.
 * `scale_colour_ptt()` and `scale_fill_ptt()` for discrete PTT colour scales
+* Complete themes `theme_ptt()` and `theme_map()`
+* `theme_ptt()` set on load.
 
 ggptt 0.1.0
 ===========