+++ NEWS.md
@@ -1,11 +1,13 @@
 ggptt 0.2.0
 ===========
-* `the_x45()` for x-axis text in 45 angle.
-* `the_legend_bot for legend.position = "bottom"`
+
 * `ptt_pal()` color PTT palette. 6 colours.
 * `scale_colour_ptt()` and `scale_fill_ptt()` for discrete PTT colour scales
 * Complete themes `theme_ptt()` and `theme_map()`
-* `theme_ptt()` set on load.
+* `theme_ptt()` and geom defaults set on load.
+* to set theme elements:
+  * `the_x45()` for x-axis text in 45 angle.
+  * `the_legend_bot` for legend.position = "bottom"
 
 ggptt 0.1.0
 ===========